+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/transition/TransitionComposeTarget.kt
@@ -17,33 +17,16 @@
 
 package com.github.panpf.sketch.transition
 
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
-import com.github.panpf.sketch.resize.ScaleDecider
-import com.github.panpf.sketch.util.fitScale
-import com.github.panpf.sketch.util.toScale
 
+/**
+ * A [Target] that supports applying [Transition]s.
+ */
+@Deprecated("Please use ComposeTarget instead and will be deleted in the future")
 interface TransitionComposeTarget : TransitionTarget {
 
     /**
      * The component's current [Painter].
      */
     val painter: Painter?
-
-    /**
-     * The [ContentScale] of the component that this target is applied to.
-     */
-    val contentScale: ContentScale
-
-    /**
-     * The [Alignment] of the component that this target is applied to.
-     */
-    val alignment: Alignment
-
-    @Deprecated("Please use contentScale instead and will be deleted in the future")
-    override val fitScale: Boolean
-        get() = contentScale.fitScale
-
-    override fun getScaleDecider(): ScaleDecider? = ScaleDecider(toScale(contentScale, alignment))
 }